+++ src/main/java/racingcargame/model/movevaluegenerator/MoveValueGenerator.java
@@ -1,4 +1,4 @@
-package racingcargame.model.carmovevalue;
+package racingcargame.model.movevaluegenerator;
 
 public interface MoveValueGenerator {
     int getMoveValue();
