+++ src/main/java/racingcargame/model/movevaluegenerator/randomnumbergenerator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcargame.model.carmovevalue.randomnumbergenerator;
+package racingcargame.model.movevaluegenerator.randomnumbergenerator;
 
 public interface NumberGenerator {
     int generateNumber(final int boundaryNumber);
