+++ src/test/java/racingcargame/utils/MoveValueGeneratorGeneratorTest.java
@@ -4,9 +4,9 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcargame.model.carmovevalue.MoveValueGenerator;
-import racingcargame.model.carmovevalue.MoveValueGeneratorByRandomNumber;
-import racingcargame.model.carmovevalue.randomnumbergenerator.NumberGenerator;
+import racingcargame.model.movevaluegenerator.MoveValueGenerator;
+import racingcargame.model.movevaluegenerator.MoveValueGeneratorByRandomNumber;
+import racingcargame.model.movevaluegenerator.randomnumbergenerator.NumberGenerator;
 
 class MoveValueGeneratorGeneratorTest {
